micro: Improve generation of data to plot the histogram
If a program allocates a lot of cell with different  sizes the histogram
would have a lot of  bars to plot, making it very hard to read.

The script to generate the histogram data is improved to take a maximum
number of cell sizes. If that limit is passed, the cell sizes are grouped
together in ranges of cell sizes, keeping the histogram readable.