Update to work with current state of the project
The naive collector is now gone, there are no statistics collection, etc.

The object files and binaries are generated in separate directories now.
The statistics graph generation is still there but not working, only
building the benchmarks works, including dil, which is added as a
submodule for simplicity (and keeping track of working version).