stats.py: Allow specifying input expression and output format
Now the script can take an arbitrary (optional) expression as the first
argument to parse the input file, using a separator (taken as the third
argument) and field specification for the expression, using $1 for the
first field, $2 for the second, etc. (similar to AWK). As second argument
the script takes the output format, in Python format specification, with
keys: min, max, mean, and std.

The defaults are '$1' for the first argument (expression to parse as
input), '%(min)s,%(mean)s,%(max)s,%(std)s' as output format and ',' as
input separator. This defaults match the old behaviour.